tweak(humans+reagents): the big scary chemical thing
PR #13097
https://github.com/ChaoticOnyx/OnyxBay/pull/13097#issue-2987234157

+++ code/modules/reagents/reagent_containers/pill.dm
@@ -32,12 +32,18 @@
 			return
 
 		to_chat(M, "<span class='notice'>You swallow \the [src].</span>")
+
+		if(ishuman(M))
+			var/mob/living/carbon/human/H = M
+			if(H.ingest(src))
+				return 1
+
 		if(reagents.total_volume)
 			reagents.trans_to_mob(M, reagents.total_volume, CHEM_INGEST)
 		qdel(src)
 		return 1
 
-	else if(istype(M, /mob/living/carbon/human))
+	else if(ishuman(M))
 		if(!M.can_force_feed(user, src))
 			return
 
@@ -49,9 +55,17 @@
 		if(!isrobot(user) && user.get_active_hand() != src)
 			return
 
+		if(!M.can_force_feed(user, src, check_resist = TRUE))
+			return
+
 		user.visible_message("<span class='warning'>[user] forces [M] to swallow \the [src].</span>")
 		var/contained = reagentlist()
 		admin_attack_log(user, M, "Fed the victim with [name] (Reagents: [contained])", "Was fed [src] (Reagents: [contained])", "used [src] (Reagents: [contained]) to feed")
+
+		var/mob/living/carbon/human/H = M
+		if(H.ingest(src))
+			return 1
+
 		if(reagents.total_volume)
 			reagents.trans_to_mob(M, reagents.total_volume, CHEM_INGEST)
 		qdel(src)
@@ -101,70 +115,70 @@
 	startswith = list(/datum/reagent/adminordrazine)
 
 /obj/item/reagent_containers/pill/stox
-	name = "Soporific (15u)"
+	name = "Soporific (15 ml)"
 	desc = "Commonly used to treat insomnia."
 	icon_state = "pill3"
 	startswith = list(/datum/reagent/soporific = 15)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/kelotane
-	name = "Kelotane (15u)"
+	name = "Kelotane (15 ml)"
 	desc = "Used to treat burns."
 	icon_state = "pill2"
 	startswith = list(/datum/reagent/kelotane = 15)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/paracetamol
-	name = "Paracetamol (15u)"
+	name = "Paracetamol (15 ml)"
 	desc = "A painkiller for the ages. Chewables!"
 	icon_state = "pill3"
 	startswith = list(/datum/reagent/painkiller/paracetamol = 15)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/tramadol
-	name = "Tramadol (15u)"
+	name = "Tramadol (15 ml)"
 	desc = "A simple painkiller."
 	icon_state = "pill3"
 	startswith = list(/datum/reagent/painkiller/tramadol = 15)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/inaprovaline
-	name = "Inaprovaline (30u)"
+	name = "Inaprovaline (30 ml)"
 	desc = "Used to stabilize patients."
 	icon_state = "pill1"
 	startswith = list(/datum/reagent/inaprovaline)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/dexalin
-	name = "Dexalin (15u)"
+	name = "Dexalin (15ml)"
 	desc = "Used to treat oxygen deprivation."
 	icon_state = "pill1"
 	startswith = list(/datum/reagent/dexalin = 15)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/dexalin_plus
-	name = "Dexalin Plus (15u)"
+	name = "Dexalin Plus (15 ml)"
 	desc = "Used to treat extreme oxygen deprivation."
 	icon_state = "pill2"
 	startswith = list(/datum/reagent/dexalinp = 15)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/dermaline
-	name = "Dermaline (15u)"
+	name = "Dermaline (15 ml)"
 	desc = "Used to treat burn wounds."
 	icon_state = "pill2"
 	startswith = list(/datum/reagent/dermaline = 15)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/dylovene
-	name = "Dylovene (15u)"
+	name = "Dylovene (15 ml)"
 	desc = "A broad-spectrum anti-toxin."
 	icon_state = "pill1"
 	startswith = list(/datum/reagent/dylovene = 15)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/bicaridine
-	name = "Bicaridine (20u)"
+	name = "Bicaridine (20 ml)"
 	desc = "Used to treat physical injuries."
 	icon_state = "pill2"
 	startswith = list(/datum/reagent/bicaridine = 20)
@@ -190,7 +204,7 @@
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/spaceacillin
-	name = "Spaceacillin (10u)"
+	name = "Spaceacillin (10 ml)"
 	desc = "Contains antiviral agents."
 	icon_state = "pill3"
 	startswith = list(/datum/reagent/spaceacillin = 10)
@@ -204,14 +218,14 @@
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/noexcutite
-	name = "Noexcutite (15u)"
+	name = "Noexcutite (15 ml)"
 	desc = "Feeling jittery? This should calm you down."
 	icon_state = "pill4"
 	startswith = list(/datum/reagent/noexcutite = 15)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/antidexafen
-	name = "Antidexafen (15u)"
+	name = "Antidexafen (15 ml)"
 	desc = "Common cold mediciation. Safe for babies!"
 	icon_state = "pill4"
 	startswith = list(
@@ -222,35 +236,35 @@
 
 //Psychiatry pills.
 /obj/item/reagent_containers/pill/methylphenidate
-	name = "Methylphenidate (15u)"
+	name = "Methylphenidate (15 ml)"
 	desc = "Improves the ability to concentrate."
 	icon_state = "pill2"
 	startswith = list(/datum/reagent/methylphenidate = 15)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/citalopram
-	name = "Citalopram (15u)"
+	name = "Citalopram (15 ml)"
 	desc = "Mild anti-depressant."
 	icon_state = "pill4"
 	startswith = list(/datum/reagent/citalopram = 15)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/paroxetine
-	name = "Paroxetine (10u)"
+	name = "Paroxetine (10 ml)"
 	desc = "Before you swallow a bullet: try swallowing this!"
 	icon_state = "pill4"
 	startswith = list(/datum/reagent/paroxetine = 10)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/hyronalin
-	name = "Hyronalin (10u)"
+	name = "Hyronalin (10 ml)"
 	desc = "Got some rads? Eat this!"
 	icon_state = "pill4"
 	startswith = list(/datum/reagent/hyronalin = 10)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/glucose
-	name = "Glucose (20u)"
+	name = "Glucose (20 ml)"
 	desc = "Used to treat blood loss"
 	icon_state = "pill4"
 	startswith = list(/datum/reagent/nutriment/glucose = 20)
@@ -259,7 +273,7 @@
 //Mining pills.
 /obj/item/reagent_containers/pill/leporazine
 	name = "Thermostabilizine"
-	desc = "Contents 15u of leporazine. Effectively stabilizes body temperature."
+	desc = "Contents 15 ml of leporazine. Effectively stabilizes body temperature."
 	icon_state = "pill2"
 	startswith = list(/datum/reagent/leporazine = 15)
 	mimic_color = TRUE
@@ -269,13 +283,13 @@
 	name = "sugar cube"
 	desc = "Sugar pressed together in block shape that is used to sweeten drinks."
 	icon_state = "sugar_cubes"
-	startswith = list(/datum/reagent/sugar = 5)
+	startswith = list(/datum/reagent/sugar = 15)
 	mimic_color = TRUE
 
 //Not actually a pill, but pills type provide everything needed for this
 /obj/item/reagent_containers/pill/cleanerpod
 	name = "space cleaner pod"
-	desc = "BLAM!-brand non-foaming space cleaner in concentrated form! Use one pod per 100u of water. Should not be consumed, but hey I'm not your mom nor a doctor."
+	desc = "BLAM!-brand non-foaming space cleaner in concentrated form! Use one pod per half a liter water. Should not be consumed, but hey I'm not your mom nor a doctor."
 	icon_state = "cleanerpod"
 	startswith = list(/datum/reagent/space_cleaner/dry = 10)
 	mimic_color = FALSE
@@ -283,56 +297,56 @@
 //Pills that probably won't be used anywhere, except in merchants or mapping, but who cares?
 
 /obj/item/reagent_containers/pill/oxycodone
-	name = "Oxycodone (15u)"
+	name = "Oxycodone (15 ml)"
 	desc = "A complex painkiller."
 	icon_state = "pill3"
 	startswith = list(/datum/reagent/painkiller/tramadol/oxycodone = 15)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/metazine
-	name = "Metazine (10u)"
+	name = "Metazine (10 ml)"
 	desc = "A combat painkiller."
 	icon_state = "pill24"
 	startswith = list(/datum/reagent/painkiller = 10)
 	mimic_color = FALSE
 
 /obj/item/reagent_containers/pill/tricordrazine
-	name = "Tricordrazine (20u)"
+	name = "Tricordrazine (20 ml)"
 	desc = "Used to slowly treat external injuries."
 	icon_state = "pill2"
 	startswith = list(/datum/reagent/tricordrazine = 20)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/alkysine
-	name = "Alkysine (5u)"
+	name = "Alkysine (5 ml)"
 	desc = "Do you have a headache? Just eat me!"
 	icon_state = "pill2"
 	startswith = list(/datum/reagent/alkysine = 5)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/imidazoline
-	name = "Imidazoline (10u)"
+	name = "Imidazoline (10 ml)"
 	desc = "Used to treat eye injuries."
 	icon_state = "pill2"
 	startswith = list(/datum/reagent/imidazoline = 10)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/ryetalyn
-	name = "Ryetalyn (5u)"
+	name = "Ryetalyn (5 ml)"
 	desc = "Used for genetic defects, including cataracts."
 	icon_state = "pill3"
 	startswith = list(/datum/reagent/ryetalyn = 5)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/peridaxon
-	name = "Peridaxon (10u)"
+	name = "Peridaxon (10 ml)"
 	desc = "Used to restore the internal organs and nervous system."
 	icon_state = "pill2"
 	startswith = list(/datum/reagent/peridaxon = 10)
 	mimic_color = TRUE
 
 /obj/item/reagent_containers/pill/albumin
-	name = "Albumin (20u)"
+	name = "Albumin (20 ml)"
 	desc = "Used to restore blood loss."
 	icon_state = "pill3"
 	startswith = list(
